levels/L2_python_env/requirements.txt
rich==13.7.1
PyYAML==6.0.2
